requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
psycopg2-binary
python-dotenv
pytest
httpx
requests
streamlit
python-multipart
